+++ docs/CHANGELOG.md
@@ -12,6 +12,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Deploy [LINSTOR Affinity Controller] as part of Piraeus Datastore.
 - Add option to set replica count for Deployment components.
 - Integration with ClusterAPI to enable graceful node evacuation in case of rolling machine updates.
+- Additional status information on `LinstorCluster`, `LinstorSatelliteConfiguration` and `LinstorSatellite` resources.
 
 [LINSTOR Affinity Controller]: https://github.com/piraeusdatastore/linstor-affinity-controller